Split ReadStore and remove service from ExternalDataset
