+++ kgforge/core/archetypes/store.py
@@ -11,35 +11,30 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with Blue Brain Nexus Forge. If not, see <https://choosealicense.com/licenses/lgpl-3.0/>.
-import json
 
 import re
-import time
-from abc import ABC, abstractmethod
+import json
+from abc import abstractmethod
 from pathlib import Path
-from typing import Any, Dict, List, Match, Optional, Tuple, Union, Type
+from typing import Any, Dict, List,  Optional, Union, Type, Match
 
 from kgforge.core import Resource
-from kgforge.core.archetypes import Mapping, Mapper
+from kgforge.core.archetypes.read_store import ReadStore
+from kgforge.core.archetypes.model import Model
+from kgforge.core.archetypes.mapping import Mapping
+from kgforge.core.archetypes.mapper import Mapper
 from kgforge.core.commons.attributes import repr_class
 from kgforge.core.commons.context import Context
 from kgforge.core.commons.exceptions import (
     DeprecationError,
-    DownloadingError,
     FreezingError,
     RegistrationError,
     TaggingError,
     UpdatingError,
     UploadingError,
-    QueryingError,
+    QueryingError
 )
 from kgforge.core.commons.execution import not_supported, run
-from kgforge.core.reshaping import collect_values
-
-# NB: Do not 'from kgforge.core.archetypes import Resolver' to avoid cyclic dependency.
-
-# FIXME: need to find a comprehensive way (different than list) to get all SPARQL reserved clauses
-from kgforge.core.wrappings.dict import DictWrapper
 
 DEFAULT_LIMIT = 100
 DEFAULT_OFFSET = 0
@@ -92,278 +87,6 @@ SPARQL_CLAUSES = [
 ]
 
 
-class ReadStore(ABC):
-
-    # See demo_store.py in kgforge/specializations/stores/ for a reference implementation.
-
-    # POLICY Methods of archetypes, except __init__, should not have optional arguments.
-
-    # POLICY Implementations should be declared in kgforge/specializations/stores/__init__.py.
-    # POLICY Implementations should not add methods but private functions in the file.
-    # TODO Move from BDD to classical testing to have a more parameterizable test suite. DKE-135.
-    # POLICY Implementations should pass tests/specializations/stores/demo_store.feature tests.
-
-    def __init__(
-            self,
-            model: Optional["Model"] = None,
-            endpoint: Optional[str] = None,
-            bucket: Optional[str] = None,
-            token: Optional[str] = None,
-            versioned_id_template: Optional[str] = None,
-            file_resource_mapping: Optional[str] = None,
-            searchendpoints: Optional[Dict] = None,
-            **store_config,
-    ) -> None:
-        # file_resource_mapping: Optional[Union[Hjson, FilePath, URL]].
-        # POLICY There could be data caching but it should be aware of changes made in the source.
-        self.model: Optional["Model"] = model
-        self.endpoint: Optional[str] = endpoint
-        self.bucket: Optional[str] = bucket
-        self.token: Optional[str] = token
-        self.versioned_id_template: Optional[str] = versioned_id_template
-        loaded = (
-            self.mapping.load(file_resource_mapping) if file_resource_mapping else None
-        )
-        self.file_mapping: Optional[Any] = loaded
-        if 'model_context' in store_config:
-            self.model_context = store_config.get('model_context')
-        elif hasattr(self.model, 'context'):
-            self.model_context = self.model.context()
-        else:
-            self.model_context = None
-        self.service: Any = self._initialize_service(
-            self.endpoint, self.bucket, self.token, searchendpoints, **store_config
-        )
-        self.context: Context = (
-            self.service.context if hasattr(self.service, "context") else None
-        )
-        self.metadata_context: Context = (
-            self.service.metadata_context
-            if hasattr(self.service, "metadata_context")
-            else None
-        )
-
-    def __repr__(self) -> str:
-        return repr_class(self)
-
-    @property
-    @abstractmethod
-    def mapping(self) -> Type[Mapping]:
-        """Mapping class to load file_resource_mapping."""
-        ...
-
-    @property
-    @abstractmethod
-    def mapper(self) -> Type[Mapper]:
-        """Mapper class to map file metadata to a Resource with file_resource_mapping."""
-        ...
-
-    # C[R]UD.
-
-    @abstractmethod
-    def retrieve(
-            self, id_: str, version: Optional[Union[int, str]], cross_bucket: bool, **params
-    ) -> Resource:
-        # POLICY Should notify of failures with exception RetrievalError including a message.
-        # POLICY Resource _store_metadata should be set using wrappers.dict.wrap_dict().
-        # POLICY Resource _synchronized should be set to True.
-        # TODO These two operations might be abstracted here when other stores will be implemented.
-        pass
-
-    def _retrieve_filename(self, id: str) -> Tuple[str, str]:
-        # TODO This operation might be adapted if other file metadata are needed.
-        not_supported()
-
-    def _prepare_download_one(
-            self,
-            url: str,
-            store_metadata: Optional[DictWrapper],
-            cross_bucket: bool
-    ) -> Tuple[str, str]:
-        # Prepare download url and download bucket
-        not_supported()
-
-    def download(
-            self,
-            data: Union[Resource, List[Resource]],
-            follow: str,
-            path: str,
-            overwrite: bool,
-            cross_bucket: bool,
-            content_type: str = None
-    ) -> None:
-        # path: DirPath.
-        urls = []
-        store_metadata = []
-        to_download = [data] if isinstance(data, Resource) else data
-        for d in to_download:
-            collected_values = collect_values(d, follow, DownloadingError)
-            urls.extend(collected_values)
-            store_metadata.extend(
-                [d._store_metadata for _ in range(len(collected_values))]
-            )
-        if len(urls) == 0:
-            raise DownloadingError(
-                f"path to follow '{follow}' was not found in any provided resource."
-            )
-        dirpath = Path(path)
-        dirpath.mkdir(parents=True, exist_ok=True)
-        timestamp = time.strftime("%Y%m%d%H%M%S")
-        filepaths = []
-        buckets = []
-        download_urls = []
-        download_store_metadata = []
-        for i, x in enumerate(urls):
-            x_download_url, x_bucket = self._prepare_download_one(x, store_metadata[i],
-                                                                  cross_bucket)
-            filename, store_content_type = self._retrieve_filename(x_download_url)
-            if not content_type or (content_type and store_content_type == content_type):
-                filepath = dirpath / filename
-                if not overwrite and filepath.exists():
-                    filepaths.append(f"{filepath}.{timestamp}")
-                else:
-                    filepaths.append(str(filepath))
-                download_urls.append(x_download_url)
-                buckets.append(x_bucket)
-                download_store_metadata.append(store_metadata[i])
-        if len(download_urls) > 1:
-            self._download_many(download_urls, filepaths, download_store_metadata, cross_bucket,
-                                content_type, buckets)
-        elif len(download_urls) == 1:
-            self._download_one(download_urls[0], filepaths[0], download_store_metadata[0],
-                               cross_bucket, content_type, buckets[0])
-        else:
-            raise DownloadingError(
-                f"No resource with content_type {content_type} was found when following the resource path '{follow}'."
-            )
-
-    def _download_many(
-            self,
-            urls: List[str],
-            paths: List[str],
-            store_metadata: Optional[List[DictWrapper]],
-            cross_bucket: bool,
-            content_type: str,
-            buckets: List[str]
-    ) -> None:
-        # paths: List[FilePath].
-        # Bulk downloading could be optimized by overriding this method in the specialization.
-        # POLICY Should follow self._download_one() policies.
-        for url, path, store_m in zip(urls, paths, store_metadata):
-            self._download_one(url, path, store_m, cross_bucket, content_type)
-
-    def _download_one(
-            self,
-            url: str,
-            path: str,
-            store_metadata: Optional[DictWrapper],
-            cross_bucket: bool,
-            content_type: str,
-            bucket: str
-    ) -> None:
-        # path: FilePath.
-        # POLICY Should notify of failures with exception DownloadingError including a message.
-        not_supported()
-
-    # Querying.
-
-    def search(
-            self, resolvers: Optional[List["Resolver"]], *filters, **params
-    ) -> List[Resource]:
-
-        # Positional arguments in 'filters' are instances of type Filter from wrappings/paths.py
-        # A dictionary can be provided for filters:
-        #  - {'key1': 'val', 'key2': {'key3': 'val'}} will be translated to
-        #  - [Filter(operator='__eq__', path=['key1'], value='val'), Filter(operator='__eq__', path=['key2', 'key3'], value='val')]
-        # Keyword arguments in 'params' could be:
-        #   - debug: bool,
-        #   - limit: int,
-        #   - offset: int,
-        #   - deprecated: bool,
-        #   - resolving: str, with values in ('exact', 'fuzzy'),
-        #   - lookup: str, with values in ('current', 'children').
-        # POLICY Should use sparql() when 'sparql' is chosen as value  for the param 'search_endpoint'.
-        # POLICY Should use elastic() when 'elastic' is chosen as value  for the param 'search_endpoint'.
-        # POLICY Given parameters for limit and offset override the input query.
-        # POLICY Should notify of failures with exception QueryingError including a message.
-        # POLICY Resource _store_metadata should be set using wrappers.dict.wrap_dict().
-        # POLICY Resource _synchronized should be set to True.
-        # TODO These two operations might be abstracted here when other stores will be implemented.
-        not_supported()
-
-    def sparql(
-            self, query: str, debug: bool, limit: int = DEFAULT_LIMIT, offset: int = DEFAULT_OFFSET,
-            **params
-    ) -> List[Resource]:
-        rewrite = params.get("rewrite", True)
-        qr = (
-            rewrite_sparql(query, self.model_context, self.service.metadata_context)
-            if self.model_context is not None and rewrite
-            else query
-        )
-        if limit:
-            qr = _replace_in_sparql(qr, "LIMIT", limit, DEFAULT_LIMIT, r" LIMIT \d+")
-        if offset:
-            qr = _replace_in_sparql(qr, "OFFSET", offset, DEFAULT_OFFSET, r" OFFSET \d+")
-        if debug:
-            self._debug_query(qr)
-        return self._sparql(qr)
-
-    def _sparql(self, query: str) -> List[Resource]:
-        # POLICY Should notify of failures with exception QueryingError including a message.
-        # POLICY Resource _store_metadata should not be set (default is None).
-        # POLICY Resource _synchronized should not be set (default is False).
-        not_supported()
-
-    def elastic(
-            self, query: str, debug: bool, limit: int = DEFAULT_LIMIT, offset: int = DEFAULT_OFFSET
-    ) -> List[Resource]:
-        query_dict = json.loads(query)
-        if limit:
-            query_dict["size"] = limit
-        if offset:
-            query_dict["from"] = offset
-        if debug:
-            self._debug_query(query_dict)
-        return self._elastic(json.dumps(query_dict))
-
-    def _elastic(self, query: str) -> List[Resource]:
-        # POLICY Should notify of failures with exception QueryingError including a message.
-        # POLICY Resource _store_metadata should not be set (default is None).
-        # POLICY Resource _synchronized should not be set (default is False).
-        not_supported()
-
-    # Versioning.
-
-    @abstractmethod
-    def _initialize_service(
-            self,
-            endpoint: Optional[str],
-            bucket: Optional[str],
-            token: Optional[str],
-            searchendpoints: Optional[Dict] = None,
-            **store_config,
-    ) -> Any:
-        # POLICY Should initialize the access to the store according to its configuration.
-        pass
-
-    @staticmethod
-    def _debug_query(query):
-        if isinstance(query, Dict):
-            print("Submitted query:", query)
-        else:
-            print(*["Submitted query:", *query.splitlines()], sep="\n   ")
-        print()
-
-    def rewrite_uri(self, uri: str, context: Context, **kwargs) -> str:
-        """Rewrite a given uri using the store Context
-        :param uri: a URI to rewrite.
-        :param context: a Store Context object
-        :return: str
-        """
-        pass
-
-
 class Store(ReadStore):
 
     # See demo_store.py in kgforge/specializations/stores/ for a reference implementation.
@@ -377,7 +100,7 @@ class Store(ReadStore):
 
     def __init__(
             self,
-            model: Optional["Model"] = None,
+            model: Optional[Model] = None,
             endpoint: Optional[str] = None,
             bucket: Optional[str] = None,
             token: Optional[str] = None,
@@ -386,9 +109,26 @@ class Store(ReadStore):
             searchendpoints: Optional[Dict] = None,
             **store_config,
     ) -> None:
-        super().__init__(model, endpoint, bucket, token, versioned_id_template,
-                         file_resource_mapping,
-                         searchendpoints, **store_config)
+        super().__init__(model)
+        self.endpoint: Optional[str] = endpoint
+        self.bucket: Optional[str] = bucket
+        self.token: Optional[str] = token
+        self.versioned_id_template: Optional[str] = versioned_id_template
+        loaded = (
+            self.mapping.load(file_resource_mapping) if file_resource_mapping else None
+        )
+        self.file_mapping: Optional[Any] = loaded
+        self.service: Any = self._initialize_service(
+            self.endpoint, self.bucket, self.token, searchendpoints, **store_config
+        )
+        self.context: Context = (
+            self.service.context if hasattr(self.service, "context") else None
+        )
+        self.metadata_context: Context = (
+            self.service.metadata_context
+            if hasattr(self.service, "metadata_context")
+            else None
+        )
 
     def __repr__(self) -> str:
         return repr_class(self)
@@ -545,6 +285,46 @@ class Store(ReadStore):
         # TODO This operation might be abstracted here when other stores will be implemented.
         not_supported()
 
+    def sparql(
+            self, query: str, debug: bool, limit: int = DEFAULT_LIMIT, offset: int = DEFAULT_OFFSET,
+            **params
+    ) -> List[Resource]:
+        rewrite = params.get("rewrite", True)
+        qr = (
+            rewrite_sparql(query, self.model_context, self.service.metadata_context)
+            if self.model_context is not None and rewrite
+            else query
+        )
+        if limit:
+            qr = _replace_in_sparql(qr, "LIMIT", limit, DEFAULT_LIMIT, r" LIMIT \d+")
+        if offset:
+            qr = _replace_in_sparql(qr, "OFFSET", offset, DEFAULT_OFFSET, r" OFFSET \d+")
+        if debug:
+            self._debug_query(qr)
+        return self._sparql(qr)
+
+    @abstractmethod
+    def _sparql():
+        ...
+
+    def elastic(
+            self, query: str, debug: bool, limit: int = DEFAULT_LIMIT, offset: int = DEFAULT_OFFSET
+    ) -> List[Resource]:
+        query_dict = json.loads(query)
+        if limit:
+            query_dict["size"] = limit
+        if offset:
+            query_dict["from"] = offset
+        if debug:
+            self._debug_query(query_dict)
+        return self._elastic(json.dumps(query_dict))
+
+    @abstractmethod
+    def _elastic(self, query: str) -> List[Resource]:
+        # POLICY Should notify of failures with exception QueryingError including a message.
+        # POLICY Resource _store_metadata should not be set (default is None).
+        # POLICY Resource _synchronized should not be set (default is False).
+        ...
     # Versioning.
 
     def freeze(self, data: Union[Resource, List[Resource]]) -> None:
